fix POST signup response
