+++ FoodInEye_app/app/src/main/java/com/example/foodineye_app/data/PostIdResponse.java
@@ -1,4 +1,4 @@
-package com.example.foodineye_app.activity;
+package com.example.foodineye_app.data;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
